+++ src/scorer/index.ts
@@ -3,12 +3,14 @@ import diversityFeedScorer from "./feed/diversityFeedScorer";
 import favsFeatureScorer from "./feature/favsFeatureScorer";
 import FeatureScorer from "./FeatureScorer";
 import FeedScorer from "./FeedScorer";
+import ImageAttachmentScorer from "./feature/ImageAttachmentScorer";
 import InteractionsFeatureScorer from "./feature/InteractionsFeatureScorer";
 import NumFavoritesScorer from "./feature/numFavoritesScorer";
 import NumRepliesScorer from "./feature/numRepliesScorer";
 import reblogsFeatureScorer from "./feature/reblogsFeatureScorer";
 import ReblogsFeedScorer from "./feed/reblogsFeedScorer";
 import TopPostFeatureScorer from "./feature/topPostFeatureScorer";
+import VideoAttachmentScorer from "./feature/VideoAttachmentScorer";
 
 
 export {
@@ -17,10 +19,12 @@ export {
     favsFeatureScorer,
     FeedScorer,
     FeatureScorer,
+    ImageAttachmentScorer,
     InteractionsFeatureScorer,
     NumFavoritesScorer,
     NumRepliesScorer,
     reblogsFeatureScorer,
     ReblogsFeedScorer,
     TopPostFeatureScorer,
+    VideoAttachmentScorer,
 };