+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+`2024-11-24`
+* Add `ImageAttachmentScorer` and `VideoAttachmentScorer`
+* Improve scorer descriptions
+
 `2024-11-23`:
 * Make time decay param configurable